Several changes

+++ opennn/config.h
@@ -64,7 +64,7 @@
 
 namespace OpenNN
 {
-    typedef float type;
+    typedef double type;
 }
